+++ src/smif/data_layer/datafile_interface.py
@@ -1,12 +1,11 @@
 """File-backed data interface
 """
 import csv
+import glob
 import os
 import re
 import shutil
 from csv import DictReader
-
-import glob
 
 import fiona
 import pyarrow as pa
@@ -413,11 +412,13 @@ class DatafileInterface(DataInterface):
 
         return data
 
-    def _read_region_names(self, region_definition_name):
-        return list(set([
+    def read_region_names(self, region_definition_name):
+        """Return the set of unique region names in region set `region_definition_name`
+        """
+        return [
             feature['properties']['name']
             for feature in self.read_region_definition_data(region_definition_name)
-        ]))
+        ]
 
     def write_region_definition(self, region_definition):
         """Write region_definition to project configuration
@@ -493,18 +494,34 @@ class DatafileInterface(DataInterface):
                 interval_definition_name))
 
         filepath = os.path.join(self.file_dir['interval_definitions'], filename)
+
+        names = {}
+
         with open(filepath, 'r') as csvfile:
             reader = DictReader(csvfile)
             data = []
-            for row in reader:
-                data.append(row)
+            for interval in reader:
+
+                name = interval['id']
+                interval_tuple = (interval['start'], interval['end'])
+                if name in names:
+                    # Append duration to existing entry
+                    self.logger.debug(
+                        "Entry %s in interval set exists at position %s", name, names[name])
+                    data[names[name]][1].append(interval_tuple)
+                else:
+                    self.logger.debug(
+                        "Add new entry %s in interval set at position %s", name, len(data))
+                    # Make a new entry
+
+                    data.append((name, [interval_tuple]))
+                    names[name] = len(data) - 1
+
         return data
 
-    def _read_interval_names(self, interval_definition_name):
-        return list(set([
-            interval['id']
-            for interval in self.read_interval_definition_data(interval_definition_name)
-        ]))
+    def read_interval_names(self, interval_definition_name):
+        return [interval[0]
+            for interval in self.read_interval_definition_data(interval_definition_name)]
 
     def write_interval_definition(self, interval_definition):
         """Write interval_definition to project configuration
@@ -837,8 +854,10 @@ class DatafileInterface(DataInterface):
             if int(datum['year']) == timestep
         ]
 
-        region_names = self._read_region_names(spatial_resolution)
-        interval_names = self._read_interval_names(temporal_resolution)
+        # Position of names in these lists dictates position of
+        # data in data array
+        region_names = self.read_region_names(spatial_resolution)
+        interval_names = self.read_interval_names(temporal_resolution)
 
         return self.data_list_to_ndarray(data, region_names, interval_names)
 
@@ -1148,8 +1167,8 @@ class DatafileInterface(DataInterface):
 
         if self.storage_format == 'local_csv':
             csv_data = self._get_data_from_csv(results_path)
-            region_names = self._read_region_names(spatial_resolution)
-            interval_names = self._read_interval_names(temporal_resolution)
+            region_names = self.read_region_names(spatial_resolution)
+            interval_names = self.read_interval_names(temporal_resolution)
             return self.data_list_to_ndarray(csv_data, region_names, interval_names)
         elif self.storage_format == 'local_binary':
             return self._get_data_from_native_file(results_path)
@@ -1183,10 +1202,9 @@ class DatafileInterface(DataInterface):
         if data.ndim == 3:
             raise NotImplementedError
         elif data.ndim == 2:
-            region_names = self._read_region_names(spatial_resolution)
-            interval_names = self._read_interval_names(temporal_resolution)
-
             if self.storage_format == 'local_csv':
+                region_names = self.read_region_names(spatial_resolution)
+                interval_names = self.read_interval_names(temporal_resolution)
                 csv_data = self.ndarray_to_data_list(data, region_names, interval_names)
                 self._write_data_to_csv(results_path, csv_data)
             elif self.storage_format == 'local_binary':
@@ -1213,63 +1231,83 @@ class DatafileInterface(DataInterface):
 
         # Return if path to previous modelruns doe snot exist
         if not os.path.isdir(results_dir):
-            self.logger.info("Warm start not possible because modelrun has no previous results (path does not exist)")
+            self.logger.info("Warm start not possible because modelrun has "
+                             "no previous results (path does not exist)")
             return None
 
         # Collect previous results
         previous_results = sorted([
-            name for name in os.listdir(results_dir) 
-            if os.path.isdir(os.path.join(results_dir, name)) 
+            name for name in os.listdir(results_dir)
+            if os.path.isdir(os.path.join(results_dir, name))
         ])
 
         # Return if no previous results exist in previous modelrun
         if len(previous_results) == 0:
-            self.logger.info("Warm start not possible because modelrun has no previous results (no results in path)")
+            self.logger.info("Warm start not possible because modelrun has "
+                             "no previous results (no results in path)")
             return None
 
-        previous_results_dir = os.path.join(self.file_dir['results'], modelrun_id, previous_results[-1])
-        results = list(glob.iglob(os.path.join(previous_results_dir, '**/*.*'), recursive=True))
+        previous_results_dir = os.path.join(self.file_dir['results'],
+                                            modelrun_id, previous_results[-1])
+        results = list(glob.iglob(os.path.join(previous_results_dir, '**/*.*'),
+                                  recursive=True))
 
         # Return if no results exist in last modelrun
         if len(results) == 0:
-            self.logger.info("Warm start not possible because there are no results in the previous modelrun")
+            self.logger.info("Warm start not possible because there are "
+                             "no results in the previous modelrun")
             return None
 
         # Return if previous results were stored in a different format
         for filename in results:
             if ((self.storage_format == 'local_csv' and not filename.endswith(".csv")) or
-            (self.storage_format == 'local_binary' and not filename.endswith(".dat"))):
-                self.logger.info("Warm start not possible because a different storage mode was used in the previous run")
+                    (self.storage_format == 'local_binary' and not filename.endswith(".dat"))):
+                self.logger.info("Warm start not possible because a different "
+                                 "storage mode was used in the previous run")
                 return None
 
         # Perform warm start
-        self.logger.info("Warm start using results from timestamp %s", previous_results[-1])
-    
+        self.logger.info("Warm start using results from timestamp %s",
+                         previous_results[-1])
+
         # Copy results from latest timestep from this modelrun_id
-        current_results_dir = os.path.join(self.file_dir['results'], modelrun_id, self.timestamp)
+        current_results_dir = os.path.join(self.file_dir['results'], modelrun_id,
+                                           self.timestamp)
         shutil.copytree(previous_results_dir, current_results_dir)
-        
+
         # Get metadata for all results
         result_metadata = []
-        for filename in glob.iglob(os.path.join(current_results_dir, '**/*.*'), recursive=True):
-            result_metadata.append(self._parse_results_path(filename.replace(self.file_dir['results'], '')[1:]))
+        for filename in glob.iglob(os.path.join(current_results_dir, '**/*.*'),
+                                   recursive=True):
+            result_metadata.append(self._parse_results_path(
+                filename.replace(self.file_dir['results'], '')[1:]))
 
         # Find latest timestep
-        result_metadata = sorted(result_metadata, key=lambda k: k['timestep'], reverse=True)
+        result_metadata = sorted(result_metadata, key=lambda k: k['timestep'],
+                                 reverse=True)
         latest_timestep = result_metadata[0]['timestep']
 
         # Remove all results with this timestep
-        results_to_remove = [result for result in result_metadata if result['timestep'] == latest_timestep]
+        results_to_remove = \
+            [result for result in result_metadata
+             if result['timestep'] == latest_timestep]
 
         for result in results_to_remove:
-            os.remove(self._get_results_path(result['modelrun_id'], result['timestamp'], result['model_name'],
-                    result['output_name'], result['spatial_resolution'], result['temporal_resolution'],
-                    result['timestep'], result['modelset_iteration'], result['decision_iteration']))
+            os.remove(self._get_results_path(result['modelrun_id'],
+                      result['timestamp'], result['model_name'],
+                      result['output_name'],
+                      result['spatial_resolution'],
+                      result['temporal_resolution'],
+                      result['timestep'],
+                      result['modelset_iteration'],
+                      result['decision_iteration']))
 
-        self.logger.info("Warm start will resume at timestep %s", latest_timestep)
+        self.logger.info("Warm start will resume at timestep %s",
+                         latest_timestep)
         return latest_timestep
 
-    def _get_results_path(self, modelrun_id, timestamp, model_name, output_name, spatial_resolution,
+    def _get_results_path(self, modelrun_id, timestamp, model_name, output_name,
+                          spatial_resolution,
                           temporal_resolution, timestep, modelset_iteration=None,
                           decision_iteration=None):
         """Return path to filename for a given output without file extension