+++ tests/convert/test_interval.py
@@ -177,9 +177,7 @@ class TestIntervalRegister():
         """Pass a time-interval definition into the register
 
         """
-        data = [{'id': '1_1',
-                 'start': 'PT0H',
-                 'end': 'PT1H'}]
+        data = [('1_1', [('PT0H', 'PT1H')])]
 
         register = TimeIntervalRegister()
         register.register(IntervalSet('energy_supply_hourly', data))
@@ -339,7 +337,7 @@ class TestValidation:
 
     def test_validate_intervals_fails(self, remap_months):
         data = remap_months
-        data.append({'id': '5', 'start': 'PT0H', 'end': 'PT1H'})
+        data.append(('5', [('PT0H', 'PT1H')]))
         with raises(ValueError) as excinfo:
             IntervalSet('remap_months', data)
         assert "Duplicate entry for hour 0 in interval set remap_months." in str(excinfo.value)